+++ CHANGELOG.md
@@ -1,10 +1,19 @@
-# [4.0.0-dev1]
+## [4.0.0-dev3]
+- Added popupManager extension on BuildContext for showing loaders
+- Added PopupManager class for showing loaders
+
+
+# [4.0.0-dev2]
+- Fixed share issue for whatsapp
+
+
+## [4.0.0-dev1]
 - Added link preview extension
 - Implemented new date extension from #60
 - Added new extension for iterable list
 
 
-# [3.5.0]
+## [3.5.0]
 
 - New version has removed all deprecated code now using.
 - Added new extension about LinkPreview for any url.
@@ -139,7 +148,7 @@
 
 ## [1.0.0] - 19/12/2020
 
-- Documantation completed.
+- Documentation completed.
 
 ## [1.1.0] - 2/1/2021